Added deletion tests and finished party creation

+++ src/main/kotlin/com/example/graphql/domain/user/UserService.kt
@@ -10,7 +10,7 @@ class UserService(private val userRepository: UserRepository) {
     }
 
     fun getAllPartyParticipants(partyId: String): List<User> {
-        return userRepository.findAllPartyParticipants(partyId.toLong())
+        return userRepository.findAllPartyParticipants(partyId)
     }
 }
